+++ noodles-sam/src/header/record/value/map/read_group/platform.rs
@@ -9,6 +9,8 @@ pub enum Platform {
     Capillary,
     /// DNBseq sequencing (`DNBSEQ`).
     DnbSeq,
+    /// Element Biosciences (`ELEMENT`).
+    Element,
     /// 454 Life Sciences sequencing (`LS454`).
     Ls454,
     /// Illumina sequencing (`ILLUMINA`).
@@ -30,6 +32,7 @@ impl AsRef<str> for Platform {
         match self {
             Self::Capillary => "CAPILLARY",
             Self::DnbSeq => "DNBSEQ",
+            Self::Element => "ELEMENT",
             Self::Ls454 => "LS454",
             Self::Illumina => "ILLUMINA",
             Self::Solid => "SOLID",
@@ -75,6 +78,7 @@ impl FromStr for Platform {
             "" => Err(ParseError::Empty),
             "CAPILLARY" => Ok(Self::Capillary),
             "DNBSEQ" => Ok(Self::DnbSeq),
+            "ELEMENT" => Ok(Self::Element),
             "LS454" => Ok(Self::Ls454),
             "ILLUMINA" => Ok(Self::Illumina),
             "SOLID" => Ok(Self::Solid),
@@ -95,6 +99,7 @@ mod tests {
     fn test_fmt() {
         assert_eq!(Platform::Capillary.to_string(), "CAPILLARY");
         assert_eq!(Platform::DnbSeq.to_string(), "DNBSEQ");
+        assert_eq!(Platform::Element.to_string(), "ELEMENT");
         assert_eq!(Platform::Ls454.to_string(), "LS454");
         assert_eq!(Platform::Illumina.to_string(), "ILLUMINA");
         assert_eq!(Platform::Solid.to_string(), "SOLID");
@@ -108,6 +113,7 @@ mod tests {
     fn test_from_str() {
         assert_eq!("CAPILLARY".parse(), Ok(Platform::Capillary));
         assert_eq!("DNBSEQ".parse(), Ok(Platform::DnbSeq));
+        assert_eq!("ELEMENT".parse(), Ok(Platform::Element));
         assert_eq!("LS454".parse(), Ok(Platform::Ls454));
         assert_eq!("ILLUMINA".parse(), Ok(Platform::Illumina));
         assert_eq!("SOLID".parse(), Ok(Platform::Solid));